+++ server-java/src/test/java/kr/hhplus/be/server/config/redis/DistributedLockAopIntegrationTest.java
@@ -4,8 +4,12 @@ import kr.hhplus.be.server.IntegrationTestSupport;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
+import org.springframework.test.context.bean.override.mockito.MockitoSpyBean;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -13,11 +17,62 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.atomic.AtomicInteger;
 
 import static org.junit.Assert.assertEquals;
+import static org.mockito.ArgumentMatchers.*;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 
-public class DistributedLockAopIntegrationTest extends IntegrationTestSupport {
+@SpringBootTest
+@ActiveProfiles("test")
+public class DistributedLockAopIntegrationTest {
+
+    @MockitoSpyBean
+    LockTemplate lockTemplate;
 
     @Autowired
     private DummyService dummyService = new DummyService();
+
+    @Test
+    @DisplayName("lock 메소드로 SIMPLE이 전달되면 LockTemplate simpleLock이 호출되어야 한다.")
+    void SIMPLE_방식으로_락을_획득하고_해제한다() throws Throwable {
+        // given
+        String param = "test";
+        String key = "test:" + param;
+
+        // when
+        dummyService.dummyMethodSimple(param);
+
+        // then
+        verify(lockTemplate, times(1)).simpleLock(eq(key), any());
+    }
+
+    @Test
+    @DisplayName("lock 메소드로 SPIN이 전달되면 LockTemplate spinLock이 호출되어야 한다.")
+    void SPIN_방식으로_락을_획득하고_해제한다() throws Throwable {
+        // given
+        String param = "test";
+        String key = "test:" + param;
+
+        // when
+        dummyService.dummyMethodSpin(param);
+
+        // then
+        verify(lockTemplate, times(1)).spinLock(eq(key), any());
+    }
+
+    @Test
+    @DisplayName("lock 메소드로 PUBSUB이 전달되면 Redisson getLock, tryLock, unlock이 호출되어야 한다.")
+    void PUBSUB_방식으로_락을_획득하고_해제한다() throws Throwable {
+        // given
+        String param = "test";
+        String key = "test:" + param;
+
+        // when
+        dummyService.dummyMethodPubSub(param);
+
+        // then
+        verify(lockTemplate, times(1)).pubSubLock(eq(key), anyLong(), anyLong(), any(), any());
+    }
 
     @Test
     @DisplayName("simple lock 방식의 동시 요청은 한 건을 제외하고 모두 실패해야 한다.")
